% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_99_ci}
\alias{calculate_99_ci}
\title{Title}
\usage{
calculate_99_ci(values)
}
\arguments{
\item{values}{numeric vector. Values used to calculate the 99 CI}
}
\value{
numeric vector
}
\description{
Title
}
