% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{calculate_grade}
\alias{calculate_grade}
\title{Title}
\usage{
calculate_grade(accuracy_val, error_info)
}
\arguments{
\item{accuracy_val}{numeric. Instrument accuracy}

\item{error_info}{dataframe}
}
\value{
dataframe
}
\description{
Title
}
