% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_accuracy_for_grading}
\alias{check_accuracy_for_grading}
\title{Title}
\usage{
check_accuracy_for_grading(sensor_chars)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}
}
\description{
Title
}
