% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_buoy_info}
\alias{check_buoy_info}
\title{Check buoy info sheet
Checks that the info sheet's format has been maintained by the user.}
\usage{
check_buoy_info(info_fpath)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}
}
\description{
Check buoy info sheet
Checks that the info sheet's format has been maintained by the user.
}
