% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_chars_error_match}
\alias{check_chars_error_match}
\title{Check error drift and sensor characteristics match
Check that the listed sensors and units match between the sensor characteristics
and error drift sheets.}
\usage{
check_chars_error_match(sensor_chars, error_drift)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\description{
Check error drift and sensor characteristics match
Check that the listed sensors and units match between the sensor characteristics
and error drift sheets.
}
