% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_error}
\alias{check_error}
\title{Title}
\usage{
check_error(error_drift, data_params_units)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}

\item{data_params_units}{dataframe. Output from read_data_params_units()}
}
\value{
printed statement if fail OR dataframe if success
}
\description{
Title
}
