% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_error_drift}
\alias{check_error_drift}
\title{Check error drift sheet.
Ensure that pre and post values match up and that there's only 1 row for each
sensor checked for each date.}
\usage{
check_error_drift(error_drift)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}
}
\description{
Check error drift sheet.
Ensure that pre and post values match up and that there's only 1 row for each
sensor checked for each date.
}
