% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_error_maint_match}
\alias{check_error_maint_match}
\title{Check dates: error drift and sensor maintenance
Check that dates match between the error drift and sensor maintenance sheets.}
\usage{
check_error_maint_match(error_drift)
}
\arguments{
\item{error_drift}{dataframe. Output from match_error_maint()}
}
\value{
print statement if error
}
\description{
Check dates: error drift and sensor maintenance
Check that dates match between the error drift and sensor maintenance sheets.
}
