% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_flags}
\alias{check_flags}
\title{Check flags
Ensure users are only supplying flags X1, X2, X3 to the tool.}
\usage{
check_flags(flags)
}
\arguments{
\item{flags}{character string. The flags set by the user in the sensor_maint sheet}
}
\value{
error message if flags were not properly assigned
}
\description{
Check flags
Ensure users are only supplying flags X1, X2, X3 to the tool.
}
