% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_missing_chunks}
\alias{check_missing_chunks}
\title{Title}
\usage{
check_missing_chunks(missing_chunks, time_small, time_large)
}
\arguments{
\item{missing_chunks}{dataframe. Edited run length encoding created in do_flag_1()}

\item{time_small}{numeric. Number of rows that counts as a small block of data}

\item{time_large}{numeric. Number of rows that counts as a large block of data}
}
\value{
dataframe
}
\description{
Title
}
