% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_params_units}
\alias{check_params_units}
\title{Check sensor characteristics match with buoy data
Check that the actual sensor names and unit names provided in the sensor
characteristics sheet match up with those in the buoy datasheet.}
\usage{
check_params_units(data_params_units, sensor_chars)
}
\arguments{
\item{data_params_units}{dataframe. Output from read_data_params_units()}

\item{sensor_chars}{dataframe. Output from read_sensor_chars()}
}
\description{
Check sensor characteristics match with buoy data
Check that the actual sensor names and unit names provided in the sensor
characteristics sheet match up with those in the buoy datasheet.
}
