% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{combine_chars_with_error}
\alias{combine_chars_with_error}
\title{Add error drift into to sensor characteristics dataframe
Enrich the sensor characteristics dataframe with the information from the error
drift sheet.}
\usage{
combine_chars_with_error(sensor_chars, error_drift)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Add error drift into to sensor characteristics dataframe
Enrich the sensor characteristics dataframe with the information from the error
drift sheet.
}
