% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_1}
\alias{do_flag_1}
\title{Title}
\usage{
do_flag_1(data_poi, time_small, time_large, repeat_0s_max)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{repeat_0s_max}{numeric. Max allowable number of repeated 0s}
}
\value{
dataframe
}
\description{
Title
}
