% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_2}
\alias{do_flag_2}
\title{Title}
\usage{
do_flag_2(data_poi, operating_range_min, operating_range_max)
}
\arguments{
\item{data_poi}{dataframe. Buoy data's datetime column, parameter of interest
and created flag columns as their respective do_flag functions are run}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}
}
\value{
dataframe
}
\description{
Title
}
