% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_3}
\alias{do_flag_3}
\title{Assign flag 3 (locally realistic range)
Assign the flag 3 if the recorded values exceed the locally realistic range
defined in the sensor characteristics sheet.}
\usage{
do_flag_3(data_poi, local_range_min, local_range_max)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{local_range_min}{numeric. Operating range minimum}

\item{local_range_max}{numeric. Operating range maximum}
}
\value{
dataframe
}
\description{
Assign flag 3 (locally realistic range)
Assign the flag 3 if the recorded values exceed the locally realistic range
defined in the sensor characteristics sheet.
}
