% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_final}
\alias{do_flag_final}
\title{Title}
\usage{
do_flag_final(data)
}
\arguments{
\item{data}{dataframe. datetime, flag_x, and all other parameter cols and
their flags}
}
\value{
dataframe
}
\description{
Title
}
