% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_m_to_4}
\alias{do_flag_m_to_4}
\title{Calculate flag for parameter of interest (not including X flags)}
\usage{
do_flag_m_to_4(data_poi)
}
\arguments{
\item{data_poi}{dataframe. Buoy data's datetime column, parameter of interest
and created flag columns as their respective do_flag functions are run}
}
\value{
dataframe
}
\description{
Calculate flag for parameter of interest (not including X flags)
}
