% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_x}
\alias{do_flag_x}
\title{Title}
\usage{
do_flag_x(data, sensor_maint)
}
\arguments{
\item{data}{dataframe. Output from read_data()}

\item{sensor_maint}{Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Title
}
