% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{edit_error_drift}
\alias{edit_error_drift}
\title{Title}
\usage{
edit_error_drift(error_drift)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Title
}
