% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{edit_plot_data}
\alias{edit_plot_data}
\title{Edit plot data only so far as it's easier to plot for R
Minor edits - data types, turning columns into factors.}
\usage{
edit_plot_data(data_poi)
}
\arguments{
\item{data_poi}{dataframe. Output from get_plot_data()}
}
\value{
ggplot
}
\description{
Edit plot data only so far as it's easier to plot for R
Minor edits - data types, turning columns into factors.
}
