% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{edit_sensor_chars}
\alias{edit_sensor_chars}
\title{Edit sensor characteristics sheet
Format the accuracy column for better read-in by the tool.}
\usage{
edit_sensor_chars(sensor_chars, error_drift)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Edit sensor characteristics sheet
Format the accuracy column for better read-in by the tool.
}
