% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{error_poi}
\alias{error_poi}
\title{Title}
\usage{
error_poi(data_poi, accuracy, error_info)
}
\arguments{
\item{data_poi}{dataframe. Output from flag_poi()}

\item{accuracy}{numeric. Instrument accuracy}

\item{error_info}{dataframe}
}
\value{
dataframe
}
\description{
Title
}
