% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Title}
\usage{
flag(data, sensor_chars, sensor_maint)
}
\arguments{
\item{data}{dataframe. Output from read_data()}

\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{sensor_maint}{Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Title
}
