% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{flag_poi}
\alias{flag_poi}
\title{Title}
\usage{
flag_poi(
  data_poi,
  operating_range_min,
  operating_range_max,
  roc_threshold,
  time_small,
  time_large,
  parameter
)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{parameter}{string. Parameter name}
}
\value{
dataframe
}
\description{
Title
}
