% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_accuracy}
\alias{format_accuracy}
\title{Title}
\usage{
format_accuracy(sca)
}
\arguments{
\item{sca}{dataframe. sensor_chars as edited in edit_sensor_chars()}
}
\value{
dataframe (or the formatted accuracy) OR NA if no accuracy was
detected OR error message if accuracy was incorrectly input
}
\description{
Title
}
