% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_datetime_string}
\alias{format_datetime_string}
\title{Title}
\usage{
format_datetime_string(datetime_format)
}
\arguments{
\item{datetime_format}{string. Datetime format as would be used by as.Date()}
}
\value{
string. The datetime format in R-readable format
}
\description{
Title
}
