% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_accuracy_units}
\alias{get_accuracy_units}
\title{Title}
\usage{
get_accuracy_units(accuracy, unit)
}
\arguments{
\item{accuracy}{string. Accuracy from sensor_chars row}

\item{unit}{string. Characters remaining after the "+/- x.xx" is extracted}
}
\value{
boolean. TRUE if accuracy input is incorrect, FALSE otherwise
}
\description{
Title
}
