% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_flag_intervals}
\alias{get_flag_intervals}
\title{Title}
\usage{
get_flag_intervals(data)
}
\arguments{
\item{data}{dataframe. Output from read_data()}
}
\value{
numeric vector
}
\description{
Title
}
