% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{handle_poi}
\alias{handle_poi}
\title{Title}
\usage{
handle_poi(input, time_small, time_large, combine_flags)
}
\arguments{
\item{input}{list. data_poi and all the sensor_chars of interest}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{combine_flags}{boolean. Should each flag be in a separate column or
should they be combined into one column}
}
\value{
dataframe
}
\description{
Title
}
