% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_plot}
\alias{make_plot}
\title{Create the plot to display in user interface
Select the colour and shape of data points depending on assigned grade, select
the background colour of the plot depending on the assigned error.}
\usage{
make_plot(data_poi)
}
\arguments{
\item{data_poi}{dataframe. Output from get_plot_data()}
}
\value{
ggplot
}
\description{
Create the plot to display in user interface
Select the colour and shape of data points depending on assigned grade, select
the background colour of the plot depending on the assigned error.
}
