% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_error_maint}
\alias{match_error_maint}
\title{Title}
\usage{
match_error_maint(error_drift, sensor_maint)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}

\item{sensor_maint}{dataframe. Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Title
}
