% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{prep_input}
\alias{prep_input}
\title{Prepare data for processing
Combine for the particular parameter of interest: the data and the sensor
characteristics.}
\usage{
prep_input(data_poi, sensor_chars_poi)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{sensor_chars_poi}{list. Sensor chars of interest associated with the
parameter of interest}
}
\value{
list
}
\description{
Prepare data for processing
Combine for the particular parameter of interest: the data and the sensor
characteristics.
}
