% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_buoy.R
\name{process_buoy}
\alias{process_buoy}
\title{Title}
\usage{
process_buoy(
  info_fpath = "data/input/example_buoy_input.xlsx",
  data_fpath = "data/input/BP_SuperBuoy_Data2022.csv",
  row_param_names = 2,
  row_data_start = 5,
  row_units = 3,
  datetime_format = "mm-dd-yyyy HH:MM:SS",
  missing_vals = c("No Data", -1e+05, ""),
  timezone = "America/Regina",
  combine_flags = TRUE
)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}

\item{data_fpath}{file path. Path to the raw buoy data file}

\item{row_param_names}{numeric. Row at which parameter names are listed}

\item{row_data_start}{numeric. Row at which buoy data begins}

\item{row_units}{numeric. Row at which parameter units are listed}

\item{datetime_format}{string. Datetime format as would be used by as.Date()}

\item{missing_vals}{string vector. Cell contents that indicate a missing value}

\item{timezone}{string. Timezone of the data used. See options with OlsonNames()}

\item{combine_flags}{boolean. TRUE results in an output with one flag column
per sensor that contains all the flags assigned. FALSE results in each
data column getting 6 associated flag columns (From flag_x to flag_4)}
}
\value{
list of dataframes output by flag_and_error()
}
\description{
Title
}
