% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{process_plot}
\alias{process_plot}
\title{Title}
\usage{
process_plot(data, index)
}
\arguments{
\item{data}{dataframe. Output from combine_buoy()}

\item{index}{numeric. List index of parameter we wish to plot}
}
\value{
ggplot
}
\description{
Title
}
