% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_params_units}
\alias{read_data_params_units}
\title{Title}
\usage{
read_data_params_units(data_fpath, row_param_names, row_units)
}
\arguments{
\item{data_fpath}{file path. Path to the raw buoy data file}

\item{row_param_names}{numeric. Row at which parameter names are listed}

\item{row_units}{numeric. Row at which parameter units are listed}
}
\value{
dataframe
}
\description{
Title
}
