% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_qcd}
\alias{read_data_qcd}
\title{Read already QCd data into the tool
Note that the QCd data must fulfill many requirements. The data format should
not be changed from what the tool outputted during the intitial QCing. Flags
and error can be changed.}
\usage{
read_data_qcd(data_qcd_fpath)
}
\arguments{
\item{data_qcd_fpath}{string. Path to QCd data csv}
}
\value{
dataframe, as was output by combine_buoy()
}
\description{
Read already QCd data into the tool
Note that the QCd data must fulfill many requirements. The data format should
not be changed from what the tool outputted during the intitial QCing. Flags
and error can be changed.
}
