% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_error_drift}
\alias{read_error_drift}
\title{Read error drift buoy info sheet
Ensure that the error drift sheet format is acceptable for read-in by
the tool.}
\usage{
read_error_drift(info_fpath, sensor_maint)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}

\item{sensor_maint}{dataframe. Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Read error drift buoy info sheet
Ensure that the error drift sheet format is acceptable for read-in by
the tool.
}
