% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sensor_chars}
\alias{read_sensor_chars}
\title{Read sensor characteristics buoy info sheet
Ensure that the sensor characteristics sheet format is acceptable for read-in by
the tool.}
\usage{
read_sensor_chars(info_fpath, error_drift)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Read sensor characteristics buoy info sheet
Ensure that the sensor characteristics sheet format is acceptable for read-in by
the tool.
}
