% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{split_data}
\alias{split_data}
\title{Unsured function}
\usage{
split_data(start_col, end_col, data_qcd)
}
\arguments{
\item{start_col}{numeric. Column at which poi data begins}

\item{end_col}{numeric. Column at which poi data ends}

\item{data_qcd}{dataframe. QCd data}
}
\value{
dataframe. The columns of interest for a specified parameter
}
\description{
Unsured function
}
