% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{split_data_qcd}
\alias{split_data_qcd}
\title{Split the already QCd data into lists
When the tool reads in the QCd data, it reads it in as a single dataframe. For
plotting, the tool must have individual lists, where each list contains all
the data relevant to a specific sensor.}
\usage{
split_data_qcd(data_qcd)
}
\arguments{
\item{data_qcd}{dataframe. Output from read_data_qcd()}
}
\value{
list of dataframes, as was output by process_buoy()
}
\description{
Split the already QCd data into lists
When the tool reads in the QCd data, it reads it in as a single dataframe. For
plotting, the tool must have individual lists, where each list contains all
the data relevant to a specific sensor.
}
