% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{calculate_grade}
\alias{calculate_grade}
\title{Calculate data grade
Calculate data grade based on how the sensor error determined during sensor
calibration and described in the error drift buoy info sheet compares to the
sensor accuracy as defined in the sensor characteristics sheet.}
\usage{
calculate_grade(sensor_accuracy, error_info)
}
\arguments{
\item{sensor_accuracy}{numeric. Instrument accuracy as output by format_accuracy()}

\item{error_info}{dataframe}
}
\value{
dataframe
}
\description{
Calculate data grade
Calculate data grade based on how the sensor error determined during sensor
calibration and described in the error drift buoy info sheet compares to the
sensor accuracy as defined in the sensor characteristics sheet.
}
