% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_input_class}
\alias{check_input_class}
\title{Title}
\usage{
check_input_class(input, input_name, classes_correct, cols_correct)
}
\arguments{
\item{input}{dataframe. The input whose column classes are to be checked}

\item{input_name}{string. The name of the input being checked}

\item{classes_correct}{character list. The correct classes for the columns of interest}

\item{cols_correct}{character list. The names of the columns to check}
}
\value{
character list. The columns that were entered incorrectly
}
\description{
Title
}
