% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_params_units}
\alias{check_params_units}
\title{Title}
\usage{
check_params_units(data_params_units, sensor_chars)
}
\arguments{
\item{data_params_units}{dataframe. Output from read_data_params_units()}

\item{sensor_chars}{dataframe. Output from read_sensor_chars()}
}
\value{
dataframe
}
\description{
Title
}
