% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_buoy.R
\name{combine_buoy}
\alias{combine_buoy}
\title{Title}
\usage{
combine_buoy(data)
}
\arguments{
\item{data}{list of dataframes. Output from process_buoy()}
}
\value{
dataframe. All buoy data in one single dataframe
}
\description{
Title
}
