% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_1}
\alias{do_flag_1}
\title{Title}
\usage{
do_flag_1(data_poi, time_small, time_large)
}
\arguments{
\item{data_poi}{dataframe. Buoy data's datetime column, parameter of interest
and created flag columns as their respective do_flag functions are run}

\item{time_small}{numeric. Number of rows that counts as a small block of data}

\item{time_large}{numeric. Number of rows that counts as a large block of data}
}
\value{
dataframe
}
\description{
Title
}
