% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_3}
\alias{do_flag_3}
\title{Title}
\usage{
do_flag_3(data_poi)
}
\arguments{
\item{data_poi}{dataframe. Buoy data's datetime column, parameter of interest
and created flag columns as their respective do_flag functions are run}
}
\value{
dataframe
}
\description{
Title
}
