% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_4}
\alias{do_flag_4}
\title{Assign flag 4 (maximum rate of change)
Assign the flag 4 if the values are defined as outliers based on the
timeseries decomposition definition of outliers OR if the rate of change
between recorded values exceeds the maximum allowable rate of change defined
in the sensor characteristics sheet.}
\usage{
do_flag_4(data_poi, roc_threshold, time_small)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{time_small}{numeric. Number of rows that indicate a small time step}
}
\value{
dataframe
}
\description{
Assign flag 4 (maximum rate of change)
Assign the flag 4 if the values are defined as outliers based on the
timeseries decomposition definition of outliers OR if the rate of change
between recorded values exceeds the maximum allowable rate of change defined
in the sensor characteristics sheet.
}
