% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_m_to_4}
\alias{do_flag_m_to_4}
\title{Calculate flag for parameter of interest (not including X flags)}
\usage{
do_flag_m_to_4(data_poi, combine_flags)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{combine_flags}{boolean. TRUE results in an output with one flag column
per sensor that contains all the flags assigned. FALSE results in each
data column getting 6 associated flag columns (From flag_x to flag_4)}
}
\value{
dataframe
}
\description{
Calculate flag for parameter of interest (not including X flags)
}
