% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{do_flag_x}
\alias{do_flag_x}
\title{Assign X flags
Assign the X flags as they are defined in the sensor maintenance sheet based
on the datetime of the flag assignment. Also assign that flag 10 minutes
ahead of and 30 minutes after the given time period to account for buoy
settling.}
\usage{
do_flag_x(data, sensor_maint)
}
\arguments{
\item{data}{dataframe. Output from read_data()}

\item{sensor_maint}{Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Assign X flags
Assign the X flags as they are defined in the sensor maintenance sheet based
on the datetime of the flag assignment. Also assign that flag 10 minutes
ahead of and 30 minutes after the given time period to account for buoy
settling.
}
