% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{edit_error_drift}
\alias{edit_error_drift}
\title{Edit error drift sheet
Calculate and store specific values required to later calculate error.}
\usage{
edit_error_drift(error_drift, sensor_maint)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}

\item{sensor_maint}{dataframe. Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Edit error drift sheet
Calculate and store specific values required to later calculate error.
}
