% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{edit_sensor_chars}
\alias{edit_sensor_chars}
\title{Title}
\usage{
edit_sensor_chars(sensor_chars)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}
}
\value{
dataframe
}
\description{
Title
}
