% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_inputs.R
\name{edit_sensor_maint}
\alias{edit_sensor_maint}
\title{Title}
\usage{
edit_sensor_maint(sensor_maint, timezone)
}
\arguments{
\item{sensor_maint}{dataframe. Output from read_sensor_maint()}

\item{timezone}{string. Timezone of the data used. See options with OlsonNames()}
}
\value{
dataframe
}
\description{
Title
}
