% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{error_poi}
\alias{error_poi}
\title{Assign error to the dataset of the parameter of interest
See the Help tab of the tool's user interface for more info on error
assignment.}
\usage{
error_poi(data_poi, sensor_accuracy, error_info, parameter)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{sensor_accuracy}{numeric. Instrument accuracy as output by format_accuracy()}

\item{error_info}{dataframe}

\item{parameter}{string. Parameter name}
}
\value{
dataframe
}
\description{
Assign error to the dataset of the parameter of interest
See the Help tab of the tool's user interface for more info on error
assignment.
}
