% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{finalize_poi}
\alias{finalize_poi}
\title{Title}
\usage{
finalize_poi(data_poi, data_flagx)
}
\arguments{
\item{data_poi}{dataframe. Output from handle_poi()}

\item{data_flagx}{dataframe. The complete datetime and flag_x columns}
}
\value{
dataframe
}
\description{
Title
}
