% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{flag_and_error}
\alias{flag_and_error}
\title{Title}
\usage{
flag_and_error(data, sensor_chars, sensor_maint, combine_flags, missing_vals)
}
\arguments{
\item{data}{dataframe. Output from read_data()}

\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{sensor_maint}{Output from read_sensor_maint()}

\item{combine_flags}{boolean. TRUE results in an output with one flag column
per sensor that contains all the flags assigned. FALSE results in each
data column getting 6 associated flag columns (From flag_x to flag_4)}

\item{missing_vals}{string vector. Cell contents that indicate a missing value}
}
\value{
dataframe
}
\description{
Title
}
