% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag_poi}
\alias{flag_poi}
\title{Title}
\usage{
flag_poi(
  data_poi,
  operating_range_min,
  operating_range_max,
  roc_threshold,
  time_small,
  time_large
)
}
\arguments{
\item{data_poi}{dataframe. Buoy data's datetime column, parameter of interest
and created flag columns as their respective do_flag functions are run}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{time_small}{numeric. Number of rows that counts as a small block of data}

\item{time_large}{numeric. Number of rows that counts as a large block of data}
}
\value{
dataframe
}
\description{
Title
}
