% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{flag_poi}
\alias{flag_poi}
\title{Title}
\usage{
flag_poi(
  data_poi,
  operating_range_min,
  operating_range_max,
  roc_threshold,
  local_range_min,
  local_range_max,
  time_small,
  time_large,
  repeat_0s_max,
  parameter,
  combine_flags = FALSE
)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{local_range_min}{numeric. Operating range minimum}

\item{local_range_max}{numeric. Operating range maximum}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{repeat_0s_max}{numeric. Max allowable number of repeated 0s}

\item{parameter}{string. Parameter name}

\item{combine_flags}{boolean. Should each flag be in a separate column or
should they be combined into one column}
}
\value{
dataframe
}
\description{
Title
}
