% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{flag_poi}
\alias{flag_poi}
\title{Assign flags to the dataset of the parameter of interest
See the Help tab of the tool's user interface for more info on flag assignment.}
\usage{
flag_poi(
  data_poi,
  operating_range_min,
  operating_range_max,
  roc_threshold,
  local_range_min,
  local_range_max,
  time_small,
  time_large,
  repeat_0s_max,
  parameter,
  combine_flags = FALSE,
  missing_vals
)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{local_range_min}{numeric. Operating range minimum}

\item{local_range_max}{numeric. Operating range maximum}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{repeat_0s_max}{numeric. Max allowable number of repeated 0s}

\item{parameter}{string. Parameter name}

\item{combine_flags}{boolean. TRUE results in an output with one flag column
per sensor that contains all the flags assigned. FALSE results in each
data column getting 6 associated flag columns (From flag_x to flag_4)}

\item{missing_vals}{string vector. Cell contents that indicate a missing value}
}
\value{
dataframe
}
\description{
Assign flags to the dataset of the parameter of interest
See the Help tab of the tool's user interface for more info on flag assignment.
}
