% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_datetime_string}
\alias{format_datetime_string}
\title{Format user-input datetime string
Formats the user-input datetime string to one that R can read}
\usage{
format_datetime_string(datetime_format)
}
\arguments{
\item{datetime_format}{string. Datetime format as would be used by as.Date()}
}
\value{
string. The datetime format in R-readable format
}
\description{
Format user-input datetime string
Formats the user-input datetime string to one that R can read
}
