% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_flag_intervals}
\alias{get_flag_intervals}
\title{Determine small vs large time periods
Depending on the time interval in the buoy datasheet, determine what defines a
small vs a large period of time (for use in flag assignment).}
\usage{
get_flag_intervals(data)
}
\arguments{
\item{data}{dataframe. Output from read_data()}
}
\value{
numeric vector
}
\description{
Determine small vs large time periods
Depending on the time interval in the buoy datasheet, determine what defines a
small vs a large period of time (for use in flag assignment).
}
