% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{handle_poi}
\alias{handle_poi}
\title{Title}
\usage{
handle_poi(
  data_poi,
  operating_range_min,
  operating_range_max,
  roc_threshold,
  accuracy,
  error_info,
  time_small,
  time_large
)
}
\arguments{
\item{data_poi}{dataframe. Datetime column and parameter of interest column}

\item{operating_range_min}{numeric. Operating range minimum}

\item{operating_range_max}{numeric. Operating range maximum}

\item{roc_threshold}{numeric. Maximum allowed rate of change}

\item{accuracy}{numeric. Instrument accuracy}

\item{error_info}{dataframe}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}
}
\value{
dataframe
}
\description{
Title
}
