% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_poi.R
\name{handle_poi}
\alias{handle_poi}
\title{Assign flags and error to the dataset of the parameter of interest}
\usage{
handle_poi(input, time_small, time_large, combine_flags, missing_vals)
}
\arguments{
\item{input}{list. data_poi and all the sensor_chars of interest}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}

\item{combine_flags}{boolean. TRUE results in an output with one flag column
per sensor that contains all the flags assigned. FALSE results in each
data column getting 6 associated flag columns (From flag_x to flag_4)}

\item{missing_vals}{string vector. Cell contents that indicate a missing value}
}
\value{
dataframe
}
\description{
Assign flags and error to the dataset of the parameter of interest
}
