% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_chars_error}
\alias{match_chars_error}
\title{Title}
\usage{
match_chars_error(sensor_chars, error_drift)
}
\arguments{
\item{sensor_chars}{dataframe. Output from read_sensor_chars()}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Title
}
