% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_error_maint}
\alias{match_error_maint}
\title{Match dates: error drift and sensor maintenance
Match the dates found in the error drift sheet to those in the sensor
maintenance sheet.}
\usage{
match_error_maint(error_drift, sensor_maint)
}
\arguments{
\item{error_drift}{dataframe. Output from read_error_drift()}

\item{sensor_maint}{dataframe. Output from read_sensor_maint()}
}
\value{
dataframe
}
\description{
Match dates: error drift and sensor maintenance
Match the dates found in the error drift sheet to those in the sensor
maintenance sheet.
}
