% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{missing_chunks_flag_1}
\alias{missing_chunks_flag_1}
\title{Title}
\usage{
missing_chunks_flag_1(missing_chunks, time_small, time_large)
}
\arguments{
\item{missing_chunks}{dataframe. Edited run length encoding created in do_flag_1()}

\item{time_small}{numeric. Number of rows that indicate a small time step}

\item{time_large}{numeric. Number of rows that indicate a large time step}
}
\value{
dataframe
}
\description{
Title
}
