% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_params_units}
\alias{read_data_params_units}
\title{Check sensor characteristics match with buoy data
Ensure that the parameter and unit combinations provided by the user in the
sensor characteristics sheet match up with the sensors and units listed in
the buoy datasheet.}
\usage{
read_data_params_units(data_fpath, row_param_names, row_units, sensor_chars)
}
\arguments{
\item{data_fpath}{file path. Path to the raw buoy data file}

\item{row_param_names}{numeric. Row at which parameter names are listed}

\item{row_units}{numeric. Row at which parameter units are listed}

\item{sensor_chars}{dataframe. Output from read_sensor_chars()}
}
\value{
dataframe
}
\description{
Check sensor characteristics match with buoy data
Ensure that the parameter and unit combinations provided by the user in the
sensor characteristics sheet match up with the sensors and units listed in
the buoy datasheet.
}
