% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_qcd}
\alias{read_data_qcd}
\title{Title}
\usage{
read_data_qcd(data_qcd_fpath)
}
\arguments{
\item{data_qcd_fpath}{string. Path to QCd data csv}
}
\value{
list of dataframes, as was output by process_buoy()
}
\description{
Title
}
