% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sensor_chars}
\alias{read_sensor_chars}
\title{Title}
\usage{
read_sensor_chars(info_fpath, error_drift)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}

\item{error_drift}{dataframe. Output from read_error_drift()}
}
\value{
dataframe
}
\description{
Title
}
