% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sensor_maint}
\alias{read_sensor_maint}
\title{Title}
\usage{
read_sensor_maint(info_fpath, timezone)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}

\item{timezone}{string. Timezone of the data used. See options with OlsonNames()}
}
\value{
dataframe
}
\description{
Title
}
