% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sensor_maint}
\alias{read_sensor_maint}
\title{Read sensor maintenance buoy info sheet
Ensure that the sensor maintenance sheet format is acceptable for read-in by
the tool.}
\usage{
read_sensor_maint(info_fpath)
}
\arguments{
\item{info_fpath}{file path. Path to the buoy info excel file}
}
\value{
dataframe
}
\description{
Read sensor maintenance buoy info sheet
Ensure that the sensor maintenance sheet format is acceptable for read-in by
the tool.
}
