% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{split_data_qcd}
\alias{split_data_qcd}
\title{Title}
\usage{
split_data_qcd(data_qcd)
}
\arguments{
\item{data_qcd}{dataframe. Output from read_data_qcd()}
}
\value{
list of dataframes, as was output by process_buoy()
}
\description{
Title
}
